Datenbankanbindung

+++ Server/src/Datenbank/Dbo/Agenda.java
@@ -10,11 +10,11 @@ import Datenbank.Annotations.Tabelle;
 @Tabelle(Name = "Agenda")
 public class Agenda {
 	@Key(Autoincrement = true)
-	@Spalte(AllowNull = false, Name = "Agenda_id", Typ = SpaltenTyp.Int)
-	public int AgendaId;
+	@Spalte(AllowNull = false, Name = "Agenda_id", Typ = SpaltenTyp.Long)
+	public long AgendaId;
 	
-	@Spalte(AllowNull = false, Name = "Meeting_id", Typ = SpaltenTyp.Int)
-	public int MeetingId;
+	@Spalte(AllowNull = false, Name = "Meeting_id", Typ = SpaltenTyp.Long)
+	public long MeetingId;
 	
 	@Spalte(AllowNull = false, Name = "Name", Typ = SpaltenTyp.Varchar, MaxLength = 128)
 	public String Name;
@@ -22,8 +22,8 @@ public class Agenda {
 	@Spalte(AllowNull = false, Name = "Diskussionszeit", Typ = SpaltenTyp.Time)
 	public Time Diskussionszeit;
 	
-	@Spalte(AllowNull = false, Name = "Sort", Typ = SpaltenTyp.Int)
-	public int Sort;
+	@Spalte(AllowNull = false, Name = "Sort", Typ = SpaltenTyp.Long)
+	public long Sort;
 	
 	@Spalte(AllowNull = false, Name = "Status", Typ = SpaltenTyp.Int)
 	public int Status;
